refactor(watch): moving watch event struct to rolldown crate
