+++ crates/rolldown/src/lib.rs
@@ -23,7 +23,10 @@ pub(crate) type SharedResolver = Arc<Resolver<OsFileSystem>>;
 pub(crate) type SharedOptions = SharedNormalizedBundlerOptions;
 
 pub use crate::{
-  bundler::Bundler, bundler_builder::BundlerBuilder, types::bundle_output::BundleOutput,
+  bundler::Bundler,
+  bundler_builder::BundlerBuilder,
+  types::bundle_output::BundleOutput,
+  watch::event::{BundleEvent, WatcherEvent},
   watcher::Watcher,
 };
