+++ crates/rolldown/src/watch/mod.rs
@@ -1,3 +1,4 @@
 pub mod emitter;
+pub mod event;
 pub mod watcher;
 pub mod watcher_task;
