+++ crates/rolldown_binding/src/types/binding_watcher_event.rs
@@ -1,15 +1,16 @@
 use napi_derive::napi;
 
 use super::binding_outputs::{BindingError, to_js_diagnostic};
+use rolldown::{BundleEvent, WatcherEvent};
 
 #[napi]
 pub struct BindingWatcherEvent {
-  inner: rolldown_common::WatcherEvent,
+  inner: WatcherEvent,
 }
 
 #[napi]
 impl BindingWatcherEvent {
-  pub fn new(inner: rolldown_common::WatcherEvent) -> Self {
+  pub fn new(inner: WatcherEvent) -> Self {
     Self { inner }
   }
 
@@ -21,7 +22,7 @@ impl BindingWatcherEvent {
   #[napi]
   pub fn watch_change_data(&self) -> BindingWatcherChangeData {
     match &self.inner {
-      rolldown_common::WatcherEvent::Change(data) => {
+      WatcherEvent::Change(data) => {
         BindingWatcherChangeData { path: data.path.to_string(), kind: data.kind.to_string() }
       }
       _ => {
@@ -33,7 +34,7 @@ impl BindingWatcherEvent {
   #[napi]
   pub fn bundle_end_data(&self) -> BindingBundleEndEventData {
     match &self.inner {
-      rolldown_common::WatcherEvent::Event(rolldown_common::BundleEvent::BundleEnd(data)) => {
+      WatcherEvent::Event(BundleEvent::BundleEnd(data)) => {
         BindingBundleEndEventData { output: data.output.to_string(), duration: data.duration }
       }
       _ => {
@@ -45,7 +46,7 @@ impl BindingWatcherEvent {
   #[napi]
   pub fn bundle_event_kind(&self) -> String {
     match &self.inner {
-      rolldown_common::WatcherEvent::Event(kind) => kind.to_string(),
+      WatcherEvent::Event(kind) => kind.to_string(),
       _ => {
         unreachable!("Expected WatcherEvent::Event")
       }
@@ -55,9 +56,10 @@ impl BindingWatcherEvent {
   #[napi]
   pub fn errors(&mut self) -> Vec<napi::Either<napi::JsError, BindingError>> {
     match &mut self.inner {
-      rolldown_common::WatcherEvent::Event(rolldown_common::BundleEvent::Error(
-        rolldown_common::OutputsDiagnostics { diagnostics, cwd },
-      )) => {
+      WatcherEvent::Event(BundleEvent::Error(rolldown_common::OutputsDiagnostics {
+        diagnostics,
+        cwd,
+      })) => {
         diagnostics.iter().map(|diagnostic| to_js_diagnostic(diagnostic, cwd.clone())).collect()
       }
       _ => {
